+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.5.7:
+* Remove the onBulkLoadComplete() and onIncrementalLoadComplete() extension
+  points in AbstractDataLoaderService as they are subsumed by onDataLoadComplete()
 * Define and implement a CacheService interface and associated implementation.
 * Start to provide mechanisms for marking change sets with pseudo etags. This
   would support client-side caching of data.